+++ tiledb/cf/netcdf_engine/_utils.py
@@ -54,7 +54,8 @@ def get_netcdf_metadata(
                 with warnings.catch_warnings():
                     warnings.warn(
                         f"Attribute '{key}' has value='{value}' that not a number. "
-                        f"Using default {key}={default} instead."
+                        f"Using default {key}={default} instead.",
+                        stacklevel=3,
                     )
                 return default
             if not np.isscalar(value):
@@ -115,10 +116,25 @@ def get_variable_values(
     return values
 
 
-def get_variable_chunks(variable: netCDF4.Variable) -> Optional[Tuple[int, ...]]:
-    """Returns the chunks from a NetCDF variable if chunked and ``None`` otherwise."""
+def get_variable_chunks(
+    variable: netCDF4.Variable, unlimited_dim_size
+) -> Optional[Tuple[int, ...]]:
+    """
+    Returns the chunks from a NetCDF variable if chunked and ``None`` otherwise.
+
+
+    If one of the dimensions has a unlimited dimension, the chunk size will be
+    reduced to the unlimited_dim_size.
+    """
     chunks = variable.chunking()
-    return None if chunks is None or chunks == "contiguous" else tuple(chunks)
+    if chunks is None or chunks == "contiguous":
+        return None
+    return tuple(
+        min(ck, dim.size if unlimited_dim_size is None else unlimited_dim_size)
+        if dim.isunlimited()
+        else ck
+        for ck, dim in zip(chunks, variable.get_dims())
+    )
 
 
 @contextmanager
@@ -177,4 +193,7 @@ def safe_set_metadata(meta, key, value):
         meta[key] = value
     except ValueError as err:  # pragma: no cover
         with warnings.catch_warnings():
-            warnings.warn(f"Failed to set metadata `{key}={value}` with error: {err}")
+            warnings.warn(
+                f"Failed to set metadata `{key}={value}` with error: {err}",
+                stacklevel=3,
+            )